# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movie App Laravel API Documentation",
        "description": "This is the API documentation for our movie app using Laravel.",
        "contact": {
            "email": "zoroloffy95@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/v1/auth/settings/user": {
            "get": {
                "tags": [
                    "Settings"
                ],
                "summary": "Get current user's settings",
                "description": "Fetch the settings of the currently authenticated user.",
                "operationId": "658209aa8a64fa3990ff182b4c007647",
                "responses": {
                    "200": {
                        "description": "Successfully fetched the user's settings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Setting ID",
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "description": "User ID associated with the setting",
                                            "type": "integer"
                                        },
                                        "is_light_mode": {
                                            "description": "Indicates if light mode is enabled",
                                            "type": "boolean"
                                        },
                                        "is_full_screen": {
                                            "description": "Indicates if full-screen mode is enabled",
                                            "type": "boolean"
                                        },
                                        "is_notifiable": {
                                            "description": "Notification preference (e.g., 'yes', 'no')",
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "description": "Creation timestamp",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Last update timestamp",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing token"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/auth/settings/toggle/{key}": {
            "patch": {
                "tags": [
                    "Settings"
                ],
                "summary": "Toggle user setting",
                "description": "Toggles a specified user setting (like light mode, fullscreen, notifications). If the setting does not exist, it returns an error.",
                "operationId": "toggleSetting",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "The key of the setting to toggle (e.g., 'is_light_mode', 'is_full_screen', 'is_notifiable')",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Setting toggled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "is_light_mode toggled successfully."
                                        },
                                        "setting": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "is_light_mode": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "is_full_screen": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "is_notifiable": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-04-17T10:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-04-17T10:10:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid setting key",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid setting key."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/auth/social-login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login using social providers like Google, Facebook, Twitter",
                "operationId": "f068bcb5d1a8e876ba3842e2b3e26ce8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "access_token",
                                    "provider"
                                ],
                                "properties": {
                                    "access_token": {
                                        "type": "string",
                                        "example": "ya29.a0AfH6SM..."
                                    },
                                    "provider": {
                                        "type": "string",
                                        "enum": [
                                            "twitter",
                                            "google",
                                            "facebook"
                                        ],
                                        "example": "google"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV..."
                                        },
                                        "refresh_token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJI..."
                                        },
                                        "user": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "John Doe",
                                                "email": "john@example.com"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "provider": [
                                                    "The selected provider is invalid."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error from provider or authentication failure",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials from provider."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}